+++ tests/index.spec.js
@@ -4,3 +4,4 @@ import './initialValue.spec';
 import './normalize';
 import './onFieldsChange.spec';
 import './forceValidate.spec';
+import './message.spec';